epic
moving stuff to other files, fixing stuff for the rewrite, etc
